Move _subset to .index
